Move exec_util module into compat
