+++ lib/fluent/plugin/exec_util.rb
@@ -14,119 +14,9 @@
 #    limitations under the License.
 #
 
-require 'msgpack'
-require 'yajl'
-
-require 'fluent/engine'
-require 'fluent/plugin'
-require 'fluent/parser'
+require 'fluent/compat/exec_util'
 
 module Fluent
-  module Plugin
-    module ExecUtil
-      SUPPORTED_FORMAT = {
-        'tsv' => :tsv,
-        'json' => :json,
-        'msgpack' => :msgpack,
-      }
-
-      class Parser
-        def initialize(on_message)
-          @on_message = on_message
-        end
-      end
-
-      class TextParserWrapperParser < Parser
-        def initialize(conf, on_message)
-          @parser = Plugin.new_parser(conf['format'])
-          @parser.configure(conf)
-          super(on_message)
-        end
-
-        def call(io)
-          io.each_line(&method(:each_line))
-        end
-
-        def each_line(line)
-          line.chomp!
-          @parser.parse(line) { |time, record|
-            @on_message.call(record, time)
-          }
-        end
-      end
-
-      class TSVParser < Parser
-        def initialize(keys, on_message)
-          @keys = keys
-          super(on_message)
-        end
-
-        def call(io)
-          io.each_line(&method(:each_line))
-        end
-
-        def each_line(line)
-          line.chomp!
-          vals = line.split("\t")
-
-          record = Hash[@keys.zip(vals)]
-
-          @on_message.call(record)
-        end
-      end
-
-      class JSONParser < Parser
-        def call(io)
-          y = Yajl::Parser.new
-          y.on_parse_complete = @on_message
-          y.parse(io)
-        end
-      end
-
-      class MessagePackParser < Parser
-        def call(io)
-          @u = Fluent::Engine.msgpack_factory.unpacker(io)
-          begin
-            @u.each(&@on_message)
-          rescue EOFError
-          end
-        end
-      end
-
-      class Formatter
-      end
-
-      class TSVFormatter < Formatter
-        def initialize(in_keys)
-          @in_keys = in_keys
-          super()
-        end
-
-        def call(record, out)
-          last = @in_keys.length-1
-          for i in 0..last
-            key = @in_keys[i]
-            out << record[key].to_s
-            out << "\t" if i != last
-          end
-          out << "\n"
-        end
-      end
-
-      class JSONFormatter < Formatter
-        def call(record, out)
-          out << Yajl.dump(record) << "\n"
-        end
-      end
-
-      class MessagePackFormatter < Formatter
-        def call(record, out)
-          record.to_msgpack(out)
-        end
-      end
-    end
-  end
-
   # obsolete
-  ExecUtil = Fluent::Plugin::ExecUtil
+  ExecUtil = Fluent::Compat::ExecUtil
 end
